requirements.txt
apache-libcloud==0.15.1
wsgiref==0.1.2
